use intellij-commons/ExecHelper (was: ExecUtilClone)
