+++ build.gradle
@@ -8,6 +8,7 @@ plugins {
 }
 
 repositories {
+    mavenLocal()
     mavenCentral()
     maven {
         url 'https://repository.jboss.org'
@@ -62,7 +63,7 @@ task integrationTest(type: Test) {
 }
 
 dependencies {
-    compile 'com.redhat.devtools.intellij:intellij-common:1.1.0'
+    compile 'com.redhat.devtools.intellij:intellij-common:1.3.0-SNAPSHOT'
     compile 'com.squareup.okhttp3:okhttp:3.12.12'
     compile group: 'org.eclipse.lsp4j', name: 'org.eclipse.lsp4j.jsonrpc', version: '0.9.0'
     compile 'org.jboss.tools.rsp.api.bundles:org.jboss.tools.rsp.api:0.22.9.Final@jar'
