+++ src/main/java/com/redhat/devtools/intellij/rsp/util/RspProcessHandler.java
@@ -1,0 +1,60 @@
+/*******************************************************************************
+ * Copyright (c) 2020 Red Hat, Inc.
+ * Distributed under license by Red Hat, Inc. All rights reserved.
+ * This program is made available under the terms of the
+ * Eclipse Public License v2.0 which accompanies this distribution,
+ * and is available at http://www.eclipse.org/legal/epl-v20.html
+ *
+ * Contributors:
+ * Red Hat, Inc. - initial API and implementation
+ ******************************************************************************/
+package com.redhat.devtools.intellij.rsp.util;
+
+import com.intellij.diagnostic.LoadingState;
+import com.intellij.execution.TaskExecutor;
+import com.intellij.execution.process.BaseOSProcessHandler;
+import com.intellij.execution.process.ProcessAdapter;
+import com.intellij.execution.process.ProcessEvent;
+import com.intellij.execution.process.ProcessWaitFor;
+import com.intellij.openapi.vfs.encoding.EncodingManager;
+import org.jetbrains.annotations.NotNull;
+
+import java.nio.charset.Charset;
+
+public class RspProcessHandler extends BaseOSProcessHandler implements TaskExecutor {
+
+        private final ProcessWaitFor myWaitFor;
+
+        public RspProcessHandler(Process process, @NotNull String presentableName, String commandLine) {
+            super(process, commandLine, LoadingState.COMPONENTS_LOADED.isOccurred() ? EncodingManager.getInstance().getDefaultConsoleEncoding() : Charset.defaultCharset());
+            myWaitFor = new ProcessWaitFor(process, this, presentableName);
+        }
+
+        @Override
+        public void startNotify() {
+            addProcessListener(new ProcessAdapter() {
+                @Override
+                public void startNotified(@NotNull ProcessEvent event) {
+                    try {
+                        myWaitFor.setTerminationCallback(integer -> notifyProcessTerminated(integer));
+                    }
+                    finally {
+                        removeProcessListener(this);
+                    }
+                }
+            });
+
+            super.startNotify();
+        }
+
+        @Override
+        protected void detachProcessImpl() {
+            destroyProcessImpl();
+        }
+
+        @Override
+        public boolean isSilentlyDestroyOnClose() {
+            return true;
+        }
+
+    }
